Added hash check in login.
